Updated the Gradle Build System.
I refreshed the gradle build system to the newest android standards (and adjusted the README accordingly).
I moved all the android application code to a separate "mobile" module,
which should make it easier to handle later, when somone eventually
 adds support for Android Wear devices (it's code will go into a "wear" module).
This structure is basically the structure that Android Studio makes after creating a new project.

Change-Id: I7c7913dc9bfa3ac99ff8e738495c0ae811c6738f