Material Design update for the main page
In this huge commit I've updated the first activity of the app with 
Material Design. I had to put all this changes in one commit because 
stuff depended on one another, so I had to do it like this, otherwise 
the build would fail. I had to edit Gradle files to move to newer APIs, 
and as a result, minimum required version is moved from Gingerbread to 
Ice Cream Sandwich, which means that it now covers ~97.4% of the market. 
I've completely replaced the ComputersActivity, with 
"NewComputersActivity", and I know that this isn't the greatest practice
(to delete the old one, and create new one with different name), but we 
can refactor that afterwards and change it back to "ComputersActivity". 
That activity has been completely remade, but everything still works, so
it's good. Other than that, I've created Material Design theme (which is
AppCompat, so it works on lower APIs), and I've introduced theme colors 
from the Material Pallette. I've used orange as an obvious primary 
color,and then blue as an accent color, because it's nicely contrasted,
and blue is also an part of the icon. I plan to add MANY more commits, 
I just added this first commit to begin with.

Change-Id: I18f783ad8a3dabe63c88236d94cc1b42955543a5