Changed computers fragment textview to button
Changed TextView to Button (with borderless style), to reflect the Material
Design change. Also, now that it's a button, we should remove <a href=...>
tags, which was a terrible idea, IMO.

EDIT: Okay, it turns out turning it into a button didn't work on API 9.
I amended it back to a TextView, but styled it a bit to look like
Material Design (at least a little). It properly works now (tested
on GB). Also, reverted the Serbian string back to using <a href>
for now.

Change-Id: I9bed7c1b3bc96aa9870408cb7fd2fb597b5a4ef8