Change FAB rotate to alpha animation
Personally I've found this FAB rotation a bit silly. Additionally,
it caused issues on older Android versions where shadows aren't
dynamic but rather are hard-coded into the view, which means that
the FAB's shadow would rotate along with the FAB.

The new animation is a kind of a pulse animation (the alpha follows
a sinusoidal path) which, although I think is imperfect, seems much
better that the current one.

Originally, I wanted to create a really cool animation where the
signal bars of the icon would rotate around the bluetooth icon
when seeking, but I had some issues with clipping, and the fact that
AnimatedVectorDrawable is supported only on API 11 and up.

Change-Id: I81d334004b4db9dde6f63cb3e9502ed606b60825