tdf#112483 fix missing slide previews
little cause, big impact, and lots time wasted looking in all the wrong
places.
Turns out that it is just a change in tint behaviour with newer versions
of android and the empty previews aren't actually empty, but are the
actual preview with white painted all over.
We're using the tinting feature to adjust the color of the icons and
text in the ActionBar, but it never was intended to apply to other
views.

Change-Id: Ibb910ed4d2006ff28925ae1ed004c2a46e768f3a