tdf#118484 fix "laser pointer" feature
give it a chance to setup the slides view before giving up.
AKA a fix for some corner case (trying to activate when no presentation
is running) broke the feature completely.
There are signals for presentation finished, so I don't see why that fix
was necessary/how to trigger the initial problem–

Change-Id: If9de4aa3a269e8b6b80b771df3d638d31072a740