Update compileSdkVersion/targetSdkVersion to 33
This is also required to meet the Google Play target
API requirement of Android 13 (API level 33) starting
August 31. [1]

Didn't see anything requiring adjustments in the list of
behavior changes for apps running on Android 13 [2]
or targeting Android 13 [3] and connecting to
Impress and remote-controlling the presentation
worked fine via both, Bluetooth and Wifi with
a LineageOS 20 (based on Android 13) device.

[1] https://developer.android.com/google/play/requirements/target-sdk
[2] https://developer.android.com/about/versions/13/behavior-changes-all
[3] https://developer.android.com/about/versions/13/behavior-changes-13

Change-Id: I7d167bc611be108b3a2e099887c5a57f4ff0a020
Reviewed-on: https://gerrit.libreoffice.org/c/impress_remote/+/155971
Tested-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>